Rename replay file names
