+++ AMBuilder
@@ -125,21 +125,24 @@ def configure_main(sdk_target, sdk, cxx):
     os.path.join(builder.sourcePath, 'src', 'kz', 'profile', 'kz_profile.cpp'),
     os.path.join(builder.sourcePath, 'src', 'kz', 'quiet', 'kz_quiet.cpp'),
     os.path.join(builder.sourcePath, 'src', 'kz', 'racing', 'kz_racing.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replayitem.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replayjumpstats.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replaybot.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replaydata.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replayevents.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replaycommands.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replayplayback.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replaywatcher.cpp'),
+
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'item.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'jumpstats.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'bot.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'data.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'events.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'commands.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'playback.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'watcher.cpp'),
     os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replaysystem.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'kz_replaycompression.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'replays', 'compression.cpp'),
+    
     os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'kz_recording.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'kz_recording_events.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'kz_recording_buffer.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'kz_recording_recorders.cpp'),
-    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'kz_recording_filewriter.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'events.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'buffer.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'recorders.cpp'),
+    os.path.join(builder.sourcePath, 'src', 'kz', 'recording', 'filewriter.cpp'),
+
     os.path.join(builder.sourcePath, 'src', 'kz', 'saveloc', 'kz_saveloc.cpp'),
     os.path.join(builder.sourcePath, 'src', 'kz', 'spec', 'kz_spec.cpp'),
     os.path.join(builder.sourcePath, 'src', 'kz', 'goto', 'kz_goto.cpp'),
