+++ src/kz/replays/commands.cpp
@@ -4,12 +4,12 @@
 #include "kz/language/kz_language.h"
 #include "kz/mode/kz_mode.h"
 #include "kz/style/kz_style.h"
-#include "kz_replaycommands.h"
-#include "kz_replaydata.h"
-#include "kz_replaybot.h"
-#include "kz_replayevents.h"
-#include "kz_replayplayback.h"
-#include "kz_replaywatcher.h"
+#include "commands.h"
+#include "data.h"
+#include "bot.h"
+#include "events.h"
+#include "playback.h"
+#include "watcher.h"
 #include "utils/uuid.h"
 #include "utils/simplecmds.h"
 #include <functional>
